requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
pymilvus>=2.4.0
pymupdf>=1.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0 